pyproject.toml
[project]
name = "crucis"
version = "0.1.0"
description = "Crucis objective-driven code training and evaluation"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "gepa>=0.1.0",
    "jinja2>=3.0",
    "json-repair>=0.30",
    "litellm>=1.64.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "radon>=6.0",
    "rich>=13.0",
]

[project.scripts]
crucis = "crucis.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "hypothesis>=6.0",
    "pytest-timeout>=2.0",
    "pytest-benchmark>=4.0",
    "ruff>=0.15.0",
    "pyink>=25.0",
]
docs = [
    "mkdocs-material>=9.0",
    "mkdocs-d2-plugin>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["crucis"]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "PLR2004"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["PLR2004"]
"conftest.py" = ["PLR2004"]

[tool.pyink]
line-length = 99
target-version = ["py312"]
pyink-indentation = 4
pyink-use-majority-quotes = true

[tool.pytest.ini_options]
testpaths = ["crucis"]
markers = [
    "llm: tests that call a real LLM (deselect with -m 'not llm')",
]
